requirements.txt
alembic==1.13.2
asyncio==3.4.3
asyncpg==0.29.0
bcrypt==4.2.0
databases[postgresql]==0.9.0
fastapi==0.111.1
greenlet==3.0.3
httpx==0.27.0
pip==24.2
pydantic==2.8.2
pydantic-settings==2.4.0
pytest==8.3.2
pytest-asyncio==0.23.8
python-dotenv==1.0.1
SQLAlchemy==2.0.31
uvicorn==0.30.3
starlette~=0.37.2